use int64_t for long

+++ simStubsGen/cpp/common.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <cstddef>
 #include <limits>
 #include <string>
@@ -144,9 +145,9 @@ static void readFromStack(sim::handle_t stackHandle, int *value, const ReadOptio
     }
 }
 
-static void readFromStack(sim::handle_t stackHandle, long long *value, const ReadOptions &rdopt = {})
+static void readFromStack(sim::handle_t stackHandle, int64_t *value, const ReadOptions &rdopt = {})
 {
-    long long v;
+    int64_t v;
     if(sim::getStackInt64Value(stackHandle, &v) == 1)
     {
         *value = v;
@@ -393,7 +394,7 @@ static void writeToStack(const int &value, sim::handle_t stackHandle, const Writ
     sim::pushInt32OntoStack(stackHandle, value);
 }
 
-static void writeToStack(const long long &value, sim::handle_t stackHandle, const WriteOptions &wropt = {})
+static void writeToStack(const int64_t &value, sim::handle_t stackHandle, const WriteOptions &wropt = {})
 {
     sim::pushInt64OntoStack(stackHandle, value);
 }
